requirements_2.7.txt
libscca_python==20181227
Flask==0.10.1
unicodecsv==0.14.1
elasticsearch==6.3.1
xmltodict==0.12.0
six==1.10.0
requests==2.21.0
lxml==3.5.0
Werkzeug==0.10.1
pymongo==2.8
celery==4.3.0
Unidecode==1.0.23
registry==0.4.2
Evtx==0.4.0
anyjson==0.3.3
booleano==1.1a2
python_dateutil==2.8.0
protobuf==3.10.0
olefile==0.46
Shapely==1.6.4.post2
sphinx_rtd_theme==0.4.3
webapp2==2.5.2
winjob==0.0.12
PyYAML==5.1.2
pytest
Flask-Celery-Helper
urllib3==1.24.1
python-registry==1.0.4
python-evtx==0.6.1
redis
gunicorn==19.4.5
